channels_tutorial/requirements.txt
beautifulsoup4==4.12.3
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django
django-bootstrap5==24.2
django-dotenv==1.4.2
django-environ==0.11.2
django-registration==3.4
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pytest==8.3.2
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
selenium==4.23.1

